Fix to add query params + function separation
