+++ lib/api.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const isValid = require('./isValid');
+
 module.exports = function buildApi(router) {
 	router.get('/getting-started', (req, res, next) => {
 		const ctx = {
@@ -29,34 +31,18 @@ module.exports = function buildApi(router) {
 		});
 	});
 
-	router.get('/whats-new/:from/:to', (req, res) => {
-		//Function to check if string is numeric (should cover undefined, "", " ", null, true, false cases)
-		const isNumeric = (input) => {
-			return (input - 0) == input && ('' + input).trim().length > 0;
-		}
-
-		//Function to check if given query param is valid (both parts are numeric strings and number after decimal point is single digit numeric string)
-		const isValid = (string) => {
-			const fp = string.split('.')[0];
-			const sp = string.split('.')[1];
-
-			if (isNumeric(fp) && isNumeric(sp) && sp.length === 1)
-				return true;
-			return false;
-		}
-
-		//Access query params
-		const from = req.params.from;
-		const to = req.params.to;
+	router.get('/whats-new', (req, res) => {
+		//Access query string
+		const from = req.query.from;
+		const to = req.query.to;
 
 		if (isValid(from) && isValid(to)) {
 			res.status(200).send(`Whats new from ${from} to ${to}`);
 		}
 		else {
 			// TODO: If format isnt matching - change below
-			res.status(400).send("Format didn't match");
+			res.status(200).send("Render whats-new general page");
 		}
-
 	});
 
 	router.all('*', (req, res) => {
